feat(cli): move one-shot agent execution into ai package
