+++ packages/ai/package.json
@@ -10,6 +10,9 @@
     "./store": "./src/store/index.ts",
     "./audio": "./src/audio/index.ts",
     "./agent": "./src/agent/index.ts",
+    "./agent/native-cli": "./src/agent/native-cli/index.ts",
+    "./agent/native-runner": "./src/agent/native-runner/index.ts",
+    "./agent/runtime": "./src/agent/runtime/index.ts",
     "./agent/sandbox": "./src/agent/sandbox/index.ts",
     "./agent/sandbox/types": "./src/agent/sandbox/types.ts",
     "./agent/sandbox/plugin": "./src/agent/sandbox/plugin.ts",
