+++ packages/ai/src/agent/index.ts
@@ -120,6 +120,15 @@ export * from "./model";
 // Routing
 export * from "./routing";
 
+// Runtime
+export * from "./runtime";
+
+// Native agent runner
+export * from "./native-runner";
+
+// Native agent CLI
+export * from "./native-cli";
+
 // Video Generation
 export * from "./video-gen";
